Consoldiate types
