+++ src/lib/Components/Bidding/Components/MaxBidPicker.tsx
@@ -2,20 +2,17 @@ import React from "react"
 import { Picker, PickerProperties } from "react-native"
 import styled from "styled-components/native"
 
-interface Bid {
-  label: string
-  value: number
-}
+import { Bid } from "../types"
 
 export interface MaxBidPickerProps extends PickerProperties {
-  bids: Bid[]
+  bids: ReadonlyArray<Bid>
 }
 
 export class MaxBidPicker extends React.Component<MaxBidPickerProps> {
   render() {
     return (
       <StyledPicker {...this.props} onValueChange={this.props.onValueChange} selectedValue={this.props.selectedValue}>
-        {this.props.bids.map((bid, index) => <Picker.Item key={index} value={index} label={bid.label} />)}
+        {this.props.bids.map((bid, index) => <Picker.Item key={index} value={index} label={bid.display} />)}
       </StyledPicker>
     )
   }
