+++ src/lib/Components/Bidding/Screens/BillingAddress.tsx
@@ -14,8 +14,7 @@ import { Button } from "../Components/Button"
 import { Container } from "../Components/Containers"
 import { Input } from "../Components/Input"
 import { Title } from "../Components/Title"
-
-import { Address } from "./ConfirmFirstTimeBid"
+import { Address } from "../types"
 
 interface BillingAddressProps {
   onSubmit?: (values: Address) => void
